refactor(torghut): rename dspy committee programs
